chore:로그아웃 버튼 추가, sidebar타입 파일추가, sidebarContent 컴포넌트 분
