+++ src/components/SideBarContent/styles.ts
@@ -1,0 +1,26 @@
+import styled from 'styled-components';
+
+export const SideBarCotentContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin-top: 40px;
+
+  div {
+    width: 200px;
+    display: flex;
+    gap: 14px;
+    cursor: pointer;
+    margin-top: 20px;
+  }
+  img {
+    width: 20px;
+  }
+`;
+export const WritePost = styled.div``;
+export const WritePostImg = styled.img``;
+export const WritePostText = styled.p``;
+export const MyPost = styled.div``;
+export const MyPostImg = styled.img``;
+export const MyPostText = styled.p``;
