+++ src/types/sideBar.ts
@@ -1,0 +1,7 @@
+export type CommonPropsType = {
+  setSidebarOpen: React.Dispatch<React.SetStateAction<boolean>>;
+};
+
+export type SideBarPropsType = CommonPropsType & {
+  sidebarOpen: boolean;
+};
